pyproject.toml
[tool.poetry]
name = "gluco"
version = "0.1.0"
description = ""
authors = ["Minglei Wang <wangmlcn@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.115.6"
uvicorn = "^0.25.0"
sqlalchemy = "^2.0.25"
pydantic = {extras = ["email"], version = "^2.5.3"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
alembic = "^1.13.1"
resend = "^2.5.1"
pillow = "^11.1.0"
numpy = "^2.2.1"
openai = "^1.59.7"
aiohttp = "^3.11.11"
pymysql = "^1.1.1"
wechatpy = "^1.8.18"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.2"
httpx = "^0.28.1"
alembic = "^1.14.0"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
